refactor(top10-9): extract ItemRowView, BinRowView, AddItemSheet into separate files
Moves three private structs out of their parent views into dedicated files under Views/Bins/Components/ and Views/Bins/Sheets/. Visibility changed from private to module-internal (default); struct bodies are byte-identical.

BinDetailView: 633->560 LOC. BinsListView: 484->444 LOC.

Tests: 592/592 pass. Build clean, 0 new warnings.
