+++ Bin Brain/Bin Brain/Views/Bins/BinDetailView.swift
@@ -558,76 +558,3 @@ struct BinDetailView: View {
     }
 }
 
-// MARK: - ItemRowView
-
-private struct ItemRowView: View {
-    let item: BinItemRecord
-
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading, spacing: 4) {
-                Text(item.name).font(.headline)
-                Text(item.category ?? "Uncategorized").font(.subheadline).foregroundStyle(.secondary)
-                if let quantity = item.quantity {
-                    Text("Qty: \(quantity, specifier: "%.0f")").font(.caption).foregroundStyle(.secondary)
-                }
-            }
-            Spacer()
-            if let confidence = item.confidence {
-                ConfidenceBadge(confidence: confidence)
-            }
-        }
-    }
-}
-
-// MARK: - AddItemSheet
-
-private struct AddItemSheet: View {
-    let binId: String
-    let apiClient: APIClient
-    let viewModel: BinDetailViewModel
-    @Binding var isPresented: Bool
-
-    @State private var name: String = ""
-    @State private var category: String = ""
-    @State private var quantityText: String = ""
-
-    var body: some View {
-        NavigationStack {
-            Form {
-                Section("Item Details") {
-                    TextField("Name", text: $name)
-                    TextField("Category (optional)", text: $category)
-                    TextField("Quantity (optional)", text: $quantityText)
-                        .keyboardType(.decimalPad)
-                }
-            }
-            .navigationTitle("Add Item")
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Cancel") { isPresented = false }
-                }
-                ToolbarItem(placement: .confirmationAction) {
-                    Button("Save") {
-                        let trimmedName = name.trimmingCharacters(in: .whitespaces)
-                        guard !trimmedName.isEmpty else { return }
-                        let trimmedCategory = category.trimmingCharacters(in: .whitespaces)
-                        let categoryValue = trimmedCategory.isEmpty ? nil : trimmedCategory
-                        let quantityValue = Double(quantityText)
-                        isPresented = false
-                        Task {
-                            await viewModel.addItem(
-                                name: trimmedName,
-                                category: categoryValue,
-                                quantity: quantityValue,
-                                binId: binId,
-                                apiClient: apiClient
-                            )
-                        }
-                    }
-                    .disabled(name.trimmingCharacters(in: .whitespaces).isEmpty)
-                }
-            }
-        }
-    }
-}
